refactor: 기존 SSO 앱을 apps/web으로 이전
루트에 있던 단일 SSO 로그인·회원가입 앱을 apps/web 워크스페이스로 이전하였습니다.
진입점·빌드 설정·테스트 환경을 web 앱 하위로 이동하였습니다.
