+++ apps/web/package.json
@@ -1,0 +1,49 @@
+{
+  "name": "@auth-econovation/web",
+  "private": true,
+  "version": "0.0.0",
+  "type": "module",
+  "scripts": {
+    "dev": "vite",
+    "build": "vite build",
+    "preview": "vite preview",
+    "type-check": "tsc -b",
+    "lint": "eslint .",
+    "test": "vitest run",
+    "test:watch": "vitest",
+    "test:run": "vitest run",
+    "test:unit": "vitest run --project unit",
+    "test:integration": "vitest run --project integration",
+    "test:coverage": "vitest run --coverage"
+  },
+  "dependencies": {
+    "@auth-econovation/api": "workspace:*",
+    "@auth-econovation/ui": "workspace:*",
+    "@tanstack/react-query": "^5.95.2",
+    "axios": "^1.14.0",
+    "react": "^19.0.0",
+    "react-dom": "^19.0.0",
+    "react-router": "^7.13.0"
+  },
+  "devDependencies": {
+    "@react-grab/mcp": "^0.1.25",
+    "@tailwindcss/vite": "^4.1.18",
+    "@testing-library/dom": "^10.4.1",
+    "@testing-library/jest-dom": "^6.9.1",
+    "@testing-library/react": "^16.3.2",
+    "@testing-library/user-event": "^14.6.1",
+    "@types/node": "^25.1.0",
+    "@types/react": "^19.0.10",
+    "@types/react-dom": "^19.0.4",
+    "@vitejs/plugin-react-swc": "^3.8.0",
+    "jsdom": "^29.1.1",
+    "msw": "^2.14.6",
+    "react-grab": "^0.1.25",
+    "tailwindcss": "^4.1.18",
+    "vite": "^6.2.0",
+    "vitest": "^4.0.18"
+  },
+  "msw": {
+    "workerDirectory": ["public"]
+  }
+}
