+++ apps/web/src/components/feature/pages/login/LoginFormSection/errorCodeMap.ts
