+++ apps/web/src/components/feature/pages/sign-up/SignUpFormSection/validateGeneration.ts
