+++ apps/web/src/components/feature/pages/sign-up/SignUpHeaderSection/index.tsx
@@ -1,0 +1,13 @@
+import { Spacing, Text } from "@auth-econovation/ui";
+
+const SignUpHeaderSection = () => {
+  return (
+    <>
+      <Text size="1">Econovation</Text>
+      <Spacing size={28} direction="vertical" />
+      <Text size="4">회원가입하고 동아리 서비스를 이용해보세요!</Text>
+    </>
+  );
+};
+
+export default SignUpHeaderSection;
