--- tsconfig.app.json
@@ -1,53 +1,0 @@
-{
-  "compilerOptions": {
-    "tsBuildInfoFile": "./node_modules/.tmp/tsconfig.app.tsbuildinfo",
-    "target": "ES2020",
-    "useDefineForClassFields": true,
-    "lib": [
-      "ES2020",
-      "DOM",
-      "DOM.Iterable"
-    ],
-    "module": "ESNext",
-    "skipLibCheck": true,
-    /* Bundler mode */
-    "moduleResolution": "bundler",
-    "allowImportingTsExtensions": true,
-    "isolatedModules": true,
-    "moduleDetection": "force",
-    "noEmit": true,
-    "jsx": "react-jsx",
-    /* Linting */
-    "strict": true,
-    "noUnusedLocals": true,
-    "noUnusedParameters": true,
-    "noFallthroughCasesInSwitch": true,
-    "noUncheckedSideEffectImports": true,
-    "paths": {
-      "@/*": [
-        "./src/*"
-      ],
-      "@assets/*": [
-        "./src/assets/*"
-      ],
-      "@shared/*": [
-        "./src/components/common/shared/*"
-      ],
-      "@entities/*": [
-        "./src/components/common/entities/*"
-      ],
-      "@pages/*": [
-        "./src/components/feature/pages/*"
-      ],
-      "@widget/*": [
-        "./src/components/feature/widget/*"
-      ],
-      "@app/*": [
-        "./src/app/*"
-      ]
-    }
-  },
-  "include": [
-    "src"
-  ]
-}
